requirements.txt
asgiref==3.9.0
colorama==0.4.6
coverage==7.9.2
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.9.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
whitenoise>=6.6